refactor: integrate reservations inside each tab (spaces & objects) and remove redundant tab
