+++ frontend/src/components/StudentReservations.tsx
@@ -1,49 +1,14 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { Objects } from "../pages/Objects/Objects";
-import { MyReservations } from "../pages/Objects/components/MyReservations";
 import { Reservations } from "../pages/Reservations/Reservations";
-import { objectsService, UserObjectReservation } from "../services/objects";
 
 export function StudentReservations() {
   const [activeTab, setActiveTab] = useState("objetos");
-  const [reservations, setReservations] = useState<UserObjectReservation[]>([]);
-  const [reservationsLoading, setReservationsLoading] = useState(false);
-  const [reservationsError, setReservationsError] = useState<string | null>(null);
   
   const tabs = [
     { id: "espacios", label: "Espacios" },
     { id: "objetos", label: "Objetos" },
-    { id: "reservas", label: "Mis Reservas" }
   ];
-
-  useEffect(() => {
-    if (activeTab === "reservas") {
-      fetchReservations();
-    }
-  }, [activeTab]);
-
-  const fetchReservations = async () => {
-    try {
-      setReservationsLoading(true);
-      setReservationsError(null);
-      const data = await objectsService.getUserObjectReservations();
-      setReservations(data);
-    } catch (err) {
-      setReservationsError(err instanceof Error ? err.message : "Error al cargar reservas");
-    } finally {
-      setReservationsLoading(false);
-    }
-  };
-
-  const handleCancelReservation = async (objectId: number, rentalId: number) => {
-    try {
-      await objectsService.cancelReservation(objectId, { rental_id: rentalId });
-      await fetchReservations(); // Refresh reservations after cancellation
-    } catch (err) {
-      console.error('Error canceling reservation:', err);
-      setReservationsError(err instanceof Error ? err.message : "Error al cancelar reserva");
-    }
-  };
   
   return (
     <div className="w-full bg-background">
@@ -79,19 +44,7 @@ export function StudentReservations() {
           
           {activeTab === "objetos" && (
             <div className="h-full">
-              <Objects onReservationSuccess={() => fetchReservations()} />
-            </div>
-          )}
-          
-          {activeTab === "reservas" && (
-            <div className="h-full p-4">
-              <MyReservations
-                reservations={reservations}
-                loading={reservationsLoading}
-                error={reservationsError}
-                onCancel={handleCancelReservation}
-                onRetry={fetchReservations}
-              />
+              <Objects />
             </div>
           )}
